Move each component into own file
